+++ system-tests/gradle-multi-module/src/test/java/com/teamscale/tia/TestwiseCoverageGradleSystemTest.java
@@ -1,6 +1,7 @@
 package com.teamscale.tia;
 
 import com.teamscale.report.testwise.model.TestwiseCoverageReport;
+import com.teamscale.test.commons.Session;
 import com.teamscale.test.commons.SystemTestUtils;
 import com.teamscale.test.commons.TeamscaleMockServer;
 import org.junit.jupiter.api.AfterEach;
@@ -31,13 +32,13 @@ public class TestwiseCoverageGradleSystemTest {
 	public void testGradleTestwiseCoverageUpload() throws Exception {
 		SystemTestUtils.runGradle("gradle-project", "clean", "tiaTests", "teamscaleReportUpload", "--no-daemon");
 
-		assertThat(teamscaleMockServer.uploadedReports).hasSize(2);
-		assertThat(teamscaleMockServer.uploadedReports).allMatch(report -> report.getPartition().equals("Unit Tests"));
+		Session session = teamscaleMockServer.getOnlySession("Unit Tests");
+		assertThat(session.getReports()).hasSize(2);
 
-		TestwiseCoverageReport unitTestReport = teamscaleMockServer.parseUploadedTestwiseCoverageReport(0);
-		assertThat(unitTestReport.tests.get(0).uniformPath).isEqualTo("org/example/Test1/test1()");
-		unitTestReport = teamscaleMockServer.parseUploadedTestwiseCoverageReport(1);
-		assertThat(unitTestReport.tests.get(0).uniformPath).isEqualTo("org/example/Test2/test2()");
+		TestwiseCoverageReport unitTestReport = session.getTestwiseCoverageReport(0);
+		assertThat(unitTestReport.tests.getFirst().uniformPath).isEqualTo("org/example/Test1/test1()");
+		unitTestReport = session.getTestwiseCoverageReport(1);
+		assertThat(unitTestReport.tests.getFirst().uniformPath).isEqualTo("org/example/Test2/test2()");
 	}
 
 }