+++ system-tests/tia-maven/maven-project/sub-project-A/src/test/java/org/example/IntegrationIT.java
@@ -1,0 +1,17 @@
+package org.example;
+
+import org.junit.jupiter.api.Test;
+import org.example.SUTA;
+
+public class IntegrationIT {
+
+    @Test
+    public void itBla() throws Exception {
+        new SUTA().bla();
+    }
+
+    @Test
+    public void itFoo() {
+        new SUTA().foo();
+    }
+}