+++ CHANGELOG.md
@@ -13,6 +13,9 @@ We use [semantic versioning](http://semver.org/):
 - [breaking] _teamscale-gradle-plugin_: Support for Gradle configuration cache (see `teamscale-gradle-plugin/MIGRATION_35.md` for how to update)
 - [breaking] _teamscale-gradle-plugin_: Bumped up minimum required Gradle version to 8.10
 
+# 34.2.5
+- [fix] Improved error messages when configuration can't be read from Teamscale 
+
 # 34.2.4
 - [fix] Convert tool could fail with `IOException` when providing nested jars